Refactored and added control.go
