+++ oviewer/reader.go
@@ -16,182 +16,14 @@ import (
 	"golang.org/x/term"
 )
 
-type controlSpecifier struct {
-	searcher Searcher
-	control  control
-	chunkNum int
-	done     chan bool
-}
-
-type control string
-
-const (
-	firstControl    = "first"
-	continueControl = "read"
-	followControl   = "follow"
-	closeControl    = "close"
-	reloadControl   = "reload"
-	loadControl     = "load"
-	searchControl   = "search"
-)
-
 const FormFeed = "\f"
-
-// ControlFile controls file read and loads in chunks.
-// ControlFile can be reloaded by file name.
-func (m *Document) ControlFile(file *os.File) error {
-	go func() error {
-		atomic.StoreInt32(&m.closed, 0)
-		r, err := m.fileReader(file)
-		if err != nil {
-			atomic.StoreInt32(&m.closed, 1)
-			log.Println(err)
-		}
-		atomic.StoreInt32(&m.eof, 0)
-		reader := bufio.NewReader(r)
-		for sc := range m.ctlCh {
-			reader, err = m.control(sc, reader)
-			if sc.control != searchControl && err != nil {
-				log.Println(sc.control, err)
-			}
-			if sc.done != nil {
-				if err != nil {
-					sc.done <- false
-				} else {
-					sc.done <- true
-				}
-				close(sc.done)
-			}
-		}
-		log.Println("close m.ctlCh")
-		return nil
-	}()
-
-	m.startControl()
-	return nil
-}
-
-func (m *Document) ControlLog() error {
-	go func() error {
-		for sc := range m.ctlCh {
-			switch sc.control {
-			case reloadControl:
-				m.reset()
-			default:
-				panic(fmt.Sprintf("unexpected %s", sc.control))
-			}
-			if sc.done != nil {
-				close(sc.done)
-			}
-		}
-		log.Println("close m.ctlCh")
-		return nil
-	}()
-	return nil
-}
-
-// controlreader is the controller for io.Reader.
-// Assuming call from Exec. reload executes the argument function.
-func (m *Document) ControlReader(r io.Reader, reload func() *bufio.Reader) error {
-	reader := bufio.NewReader(r)
-	go func() error {
-		var err error
-		for sc := range m.ctlCh {
-			switch sc.control {
-			case firstControl:
-				reader, err = m.firstRead(reader)
-				if !m.BufEOF() {
-					m.continueControl()
-				}
-			case continueControl:
-				reader, err = m.continueRead(reader)
-				if !m.BufEOF() {
-					m.continueControl()
-				}
-			case reloadControl:
-				log.Println("reset")
-				reader = reload()
-				m.startControl()
-			default:
-				panic(fmt.Sprintf("unexpected %s", sc.control))
-			}
-			if err != nil {
-				log.Println(err)
-			}
-			if sc.done != nil {
-				close(sc.done)
-			}
-		}
-		log.Println("close ctlCh")
-		return nil
-	}()
-
-	m.startControl()
-	return nil
-}
-
-func (m *Document) control(sc controlSpecifier, reader *bufio.Reader) (*bufio.Reader, error) {
-	if atomic.LoadInt32(&m.closed) == 1 && sc.control != reloadControl {
-		return nil, fmt.Errorf("closed %s", sc.control)
-	}
-
-	var err error
-	switch sc.control {
-	case firstControl:
-		reader, err = m.firstRead(reader)
-		if !m.BufEOF() {
-			m.continueControl()
-		}
-	case continueControl:
-		reader, err = m.continueRead(reader)
-		if !m.BufEOF() {
-			m.continueControl()
-		}
-	case followControl:
-		reader, err = m.followRead(reader)
-	case loadControl:
-		reader, err = m.readChunk(reader, sc.chunkNum)
-	case searchControl:
-		_, err = m.searchChunk(sc.chunkNum, sc.searcher)
-		if err != nil {
-			return reader, err
-		}
-		return m.readChunk(reader, sc.chunkNum)
-	case reloadControl:
-		reader, err = m.reloadRead(reader)
-		m.startControl()
-	case closeControl:
-		err = m.close()
-		log.Println(err)
-		err = nil
-	default:
-		panic(fmt.Sprintf("unexpected %s", sc.control))
-	}
-	return reader, nil
-}
-
-func (m *Document) startControl() {
-	go func() {
-		m.ctlCh <- controlSpecifier{
-			control: firstControl,
-		}
-	}()
-}
-
-func (m *Document) continueControl() {
-	go func() {
-		m.ctlCh <- controlSpecifier{
-			control: continueControl,
-		}
-	}()
-}
 
 // firstRead first reads the file.
 // Packs the contents of the read file into the first chunk.
 func (m *Document) firstRead(reader *bufio.Reader) (*bufio.Reader, error) {
 	chunk := m.chunks[0]
-	start := 0
-	if err := m.packChunk(chunk, reader, start, true); err != nil {
+	err := m.fillChunk(chunk, reader, 0, true)
+	if err != nil {
 		if !errors.Is(err, io.EOF) {
 			atomic.StoreInt32(&m.eof, 1)
 			return nil, err
@@ -214,7 +46,7 @@ func (m *Document) continueRead(reader *bufio.Reader) (*bufio.Reader, error) {
 
 	chunk := m.lastChunk()
 	start := len(chunk.lines)
-	if err := m.readOrCountChunk(chunk, reader, start); err != nil {
+	if err := m.addChunk(chunk, reader, start); err != nil {
 		if !errors.Is(err, io.EOF) {
 			return nil, err
 		}
@@ -259,45 +91,13 @@ func (m *Document) readChunk(reader *bufio.Reader, chunkNum int) (*bufio.Reader,
 	}
 	reader.Reset(m.file)
 	start := 0
-	if err := m.packChunk(chunk, reader, start, false); err != nil {
+	if err := m.fillChunk(chunk, reader, start, false); err != nil {
 		if !errors.Is(err, io.EOF) {
 			return nil, err
 		}
 		reader = m.afterEOF(reader)
 	}
 	return reader, nil
-}
-
-func (m *Document) searchChunk(chunkNum int, searcher Searcher) (int, error) {
-	chunk := m.chunks[chunkNum]
-	if _, err := m.file.Seek(chunk.start, io.SeekStart); err != nil {
-		return 0, err
-	}
-	reader := bufio.NewReader(m.file)
-	var line bytes.Buffer
-	var isPrefix bool
-	i := 0
-	for i < ChunkSize {
-		buf, err := reader.ReadSlice('\n')
-		if err == bufio.ErrBufferFull {
-			isPrefix = true
-			err = nil
-		}
-		line.Write(buf)
-		if isPrefix {
-			isPrefix = false
-			continue
-		}
-		if searcher.Match(line.Bytes()) {
-			return i, nil
-		}
-		i++
-		line.Reset()
-		if err != nil {
-			break
-		}
-	}
-	return 0, ErrNotFound
 }
 
 // reloadRead performs reload processing
@@ -316,11 +116,12 @@ func (m *Document) reloadRead(reader *bufio.Reader) (*bufio.Reader, error) {
 	return reader, nil
 }
 
-func (m *Document) readOrCountChunk(chunk *chunk, reader *bufio.Reader, start int) error {
+// addChunk fills or reserves a chunk.
+func (m *Document) addChunk(chunk *chunk, reader *bufio.Reader, start int) error {
 	if !m.seekable {
-		return m.packChunk(chunk, reader, start, true)
+		return m.fillChunk(chunk, reader, start, true)
 	}
-	return m.countChunk(chunk, reader, start)
+	return m.reserveChunk(chunk, reader, start)
 }
 
 func (m *Document) reloadFile(reader *bufio.Reader) (*bufio.Reader, error) {
@@ -411,7 +212,7 @@ func (m *Document) readAll(reader *bufio.Reader) error {
 	chunk := m.lastChunk()
 	start := len(chunk.lines)
 	for {
-		if err := m.packChunk(chunk, reader, start, true); err != nil {
+		if err := m.fillChunk(chunk, reader, start, true); err != nil {
 			return err
 		}
 		chunk = NewChunk(m.size)
@@ -422,8 +223,8 @@ func (m *Document) readAll(reader *bufio.Reader) error {
 	}
 }
 
-// packChunk append lines read from reader into chunks.
-func (m *Document) packChunk(chunk *chunk, reader *bufio.Reader, start int, isCount bool) error {
+// fillChunk append lines read from reader into chunks.
+func (m *Document) fillChunk(chunk *chunk, reader *bufio.Reader, start int, isCount bool) error {
 	var line bytes.Buffer
 	var isPrefix bool
 	i := start
@@ -466,7 +267,8 @@ func (m *Document) packChunk(chunk *chunk, reader *bufio.Reader, start int, isCo
 	return nil
 }
 
-func (m *Document) countChunk(chunk *chunk, reader *bufio.Reader, start int) error {
+// reserveChunk reserves ChunkSize lines.
+func (m *Document) reserveChunk(chunk *chunk, reader *bufio.Reader, start int) error {
 	var isPrefix bool
 	i := start
 	for i < ChunkSize {
